varregion: remove serialization of region terrain to floats when sending patches. This should eliminate much memory thrashing and CPU usage while sending initial terrain.
The old way of passing terrain was to convert it to an array of floats. This is
really bad for large terrain (think 4096x4096 floats). This change passes a dummy
float array since the real region info is used anyway and the floats are ignored.
(The ignoring the terrain floats is a kludge so as to not change IClientAPI.)