Donating the Avination Mute Module
This lived out-of-tree and therefore wasn't part of the big code
drop. It's classic Avination, tied to MySQL and a direct database
connection and I don't have the bandwidth to fix it to core methods.
However, it has all the strange mute logic that OpenSim was missing
and maybe someone can step up to make it work in an open grid context.